Hangman Game
This project is based on a game called HANGMAN.
It is basically a word guessing game . The user has 6 options to choose from . The word to be guessed could be the name of a country, Hollywood star, Bollywood star, Hollywood movie, Bollywood movie or a sport. There are 3 levels to choose from on basis of its difficulty. The user would be given 5 chances to guess a word, thus with every wrong guess the chances of guessing that word would be decreased.